+++ Source/SysPSP/UI/AdvancedOptionsScreen.cpp
@@ -143,13 +143,13 @@ void	IAdvancedOptionsScreen::Render()
 {
 	mpContext->ClearBackground();
 
-	u32		font_height( mpContext->GetFontHeight() );
-	u32		line_height( font_height + 2 );
-	s32		y;
+	s16		font_height( mpContext->GetFontHeight() );
+	s16		line_height( font_height + 2 );
+	s16		y;
 
 	const char * const title_text = "Advanced Options";
 	mpContext->SetFontStyle( CUIContext::FS_HEADING );
-	u32		heading_height( mpContext->GetFontHeight() );
+	s16		heading_height( mpContext->GetFontHeight() );
 	y = BELOW_MENU_MIN + heading_height;
 	mpContext->DrawTextAlign( LIST_TEXT_LEFT, LIST_TEXT_WIDTH, AT_CENTRE, y, title_text, mpContext->GetDefaultTextColour() ); y += heading_height;
 	mpContext->SetFontStyle( CUIContext::FS_REGULAR );
